requirements/base.txt
Django==4.1.9
djangorestframework==3.12.4
djangorestframework-simplejwt
django-filter==23.2
django-cors-headers==4.0.0
psycopg2-binary==2.9.6

asgiref==3.7.2
backcall==0.1.0
decorator==4.4.2

ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.5

ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==2.7.0
pytz==2019.3
six==1.16.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
cryptography
exponent_server_sdk
python-dateutil
